pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "stock-market-analyzer-mcp"
version = "1.0.0"
description = "MCP server for real-time stock market data analysis"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "mcp>=0.9.0",
    "requests>=2.31.0",
    "pandas>=2.0.0",
    "python-dotenv>=1.0.0",
    "yfinance>=0.2.0"
]

[project.scripts]
stock-market-analyzer-mcp = "mcp_server:main"
